Update patch set 1
Patch Set 1:

Not sure all those flags are still used in Q or needed for this in general. Media extensions we don't set on 845 (LOS does though).

Is everything working or is there still work to do on this?

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>